This small modification was made to correct a small error: the task was looking for a style.scss file, while the only .scss file in the folder was styles.scss.
With this correction, the style task will look for any .scss file directly in the sass/ folder (the content of base/ or any other subfolder won't be rendered, unless they are imported in an other .scss file at root of the sass/ folder)
I also added a method to catch errors. It will display them in the command line window.